Widen cartography and score types
Cartography types (Prominence, MinZoom, MaxZoom, SortKey)
from uint8 to int32; ConfidenceScore from float32 to
float64. Matches actual data ranges in Overture releases.

+++ packages/overture-schema-core/src/overture/schema/core/types.py
@@ -4,12 +4,12 @@ from pydantic import (
     Field,
 )
 
-from overture.schema.system.primitive import float32, int32
+from overture.schema.system.primitive import float64, int32
 
 ConfidenceScore = NewType(
     "ConfidenceScore",
     Annotated[
-        float32,
+        float64,
         Field(description="Confidence score between 0.0 and 1.0", ge=0.0, le=1.0),
     ],
 )
